+++ lib/src/utils/media_upload/resource/media_content_type.dart
@@ -4,7 +4,8 @@ enum MediaContentType {
   jpg("image/jpg"),
   pdf("application/pdf"),
   webp("image/webp"),
-  mp4("video/mp4");
+  mp4("video/mp4"),
+  txt("text/plain"),;
 
   final String value;